+++ backend/src/db/impls/nlp_analysis.rs
@@ -1,5 +1,5 @@
 use crate::db::db_stored::DbStoredDependentlyInner;
-use crate::db::impls::report::FromDb;
+use crate::db::from_db::FromDb;
 use crate::db::model::{DbNlpAnalysis, DbNlpMetadata};
 use crate::nlp::analysis::NlpAnalysisKind;
 use crate::nlp::nlp_response::{NlpAnalysis, NlpMetadata};
