+++ backend/src/db/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod db_client;
 pub(crate) mod db_stored;
+mod from_db;
 mod impls;
 mod model;
 mod pagination;
